+++ lib/models/groq_api.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:groq_sdk/extensions/groq_json_extensions.dart';
 import 'package:groq_sdk/models/groq_chat.dart';
 import 'package:groq_sdk/models/groq_conversation_item.dart';
 import 'package:groq_sdk/models/groq_exceptions.dart';
@@ -7,36 +8,46 @@ import 'package:groq_sdk/models/groq_llm_model.dart';
 import 'package:groq_sdk/models/groq_message.dart';
 import 'package:groq_sdk/models/groq_rate_limit_information.dart';
 import 'package:groq_sdk/models/groq_response.dart';
+import 'package:groq_sdk/models/groq_audio_response.dart';
 import 'package:groq_sdk/models/groq_usage.dart';
 import 'package:groq_sdk/utils/auth_http.dart';
+import 'package:groq_sdk/utils/groq_parser.dart';
+import 'package:http/http.dart' as http;
 
 class GroqApi {
   static const String _chatCompletionUrl =
       'https://api.groq.com/openai/v1/chat/completions';
   static const String _getModelBaseUrl =
       'https://api.groq.com/openai/v1/models';
+  static const String _getAudioTranscriptionUrl =
+      'https://api.groq.com/openai/v1/audio/transcriptions';
+  static const String _getAudioTranslationUrl =
+      'https://api.groq.com/openai/v1/audio/translations';
 
+  ///Returns the model metadata from groq with the given model id
   static Future<GroqLLMModel> getModel(String modelId, String apiKey) async {
     final response =
         await AuthHttp.get(url: '$_getModelBaseUrl/$modelId', apiKey: apiKey);
     if (response.statusCode == 200) {
-      return GroqLLMModel.fromJson(json.decode(response.body));
+      return GroqParser.llmModelFromJson(json.decode(response.body));
     } else {
       throw GroqException.fromResponse(response);
     }
   }
 
+  ///Returns a list of all model metadatas available in Groq
   static Future<List<GroqLLMModel>> listModels(String apiKey) async {
     final response = await AuthHttp.get(url: _getModelBaseUrl, apiKey: apiKey);
     if (response.statusCode == 200) {
       final Map<String, dynamic> jsonData = json.decode(response.body);
       final List<dynamic> jsonList = jsonData['data'] as List<dynamic>;
-      return jsonList.map((json) => GroqLLMModel.fromJson(json)).toList();
+      return jsonList.map((json) => GroqParser.llmModelFromJson(json)).toList();
     } else {
       throw GroqException.fromResponse(response);
     }
   }
 
+  ///Returns a new chat instance with the given model id
   static Future<(GroqResponse, GroqUsage, GroqRateLimitInformation)>
       getNewChatCompletion({
     required String apiKey,
@@ -62,11 +73,13 @@ class GroqApi {
         url: _chatCompletionUrl, apiKey: apiKey, body: jsonMap);
     //Rate Limit information
     final rateLimitInfo =
-        GroqRateLimitInformation.fromHeaders(response.headers);
+        GroqParser.rateLimitInformationFromHeaders(response.headers);
     if (response.statusCode < 300) {
       final Map<String, dynamic> jsonData = json.decode(response.body);
-      final GroqResponse groqResponse = GroqResponse.fromJson(jsonData);
-      final GroqUsage groqUsage = GroqUsage.fromJson(jsonData["usage"]);
+      final GroqResponse groqResponse =
+          GroqParser.groqResponseFromJson(jsonData);
+      final GroqUsage groqUsage =
+          GroqParser.groqUsageFromChatJson(jsonData["usage"]);
       return (groqResponse, groqUsage, rateLimitInfo);
     } else if (response.statusCode == 429) {
       throw GroqRateLimitException(
@@ -76,6 +89,72 @@ class GroqApi {
       );
     } else {
       throw GroqException.fromResponse(response);
+    }
+  }
+
+  ///transcribes the audio file at the given path using the model with the given model id
+  static Future<(GroqAudioResponse, GroqUsage, GroqRateLimitInformation)>
+      transcribeAudio({
+    required String apiKey,
+    required String filePath,
+    required String modelId,
+  }) async {
+    var request =
+        http.MultipartRequest('POST', Uri.parse(_getAudioTranscriptionUrl));
+
+    request.headers['Authorization'] = 'Bearer $apiKey';
+    request.headers['Content-Type'] = 'multipart/form-data';
+    request.files.add(await http.MultipartFile.fromPath('file', filePath));
+    request.fields['model'] = modelId;
+
+    var response = await request.send();
+    final responseBody = await response.stream.bytesToString();
+
+    final jsonBody = json.decode(responseBody);
+    if (response.statusCode == 200) {
+      final audioResponse = GroqParser.audioResponseFromJson(jsonBody);
+      final usage =
+          GroqParser.groqUsageFromAudioJson(jsonBody['x_groq']['usage']);
+      final rateLimitInfo =
+          GroqParser.rateLimitInformationFromHeaders(response.headers);
+      return (audioResponse, usage, rateLimitInfo);
+    } else {
+      throw GroqException(
+          statusCode: response.statusCode, error: GroqError.fromJson(jsonBody));
+    }
+  }
+
+  ///Translates the audio file at the given file path to text
+  static Future<(GroqAudioResponse, GroqUsage, GroqRateLimitInformation)>
+      translateAudio({
+    required String apiKey,
+    required String filePath,
+    required String modelId,
+    required double temperature,
+  }) async {
+    var request =
+        http.MultipartRequest('POST', Uri.parse(_getAudioTranslationUrl));
+
+    request.headers['Authorization'] = 'Bearer $apiKey';
+    request.headers['Content-Type'] = 'multipart/form-data';
+    request.files.add(await http.MultipartFile.fromPath('file', filePath));
+    request.fields['model'] = modelId;
+    request.fields['temperature'] = temperature.toString();
+
+    var response = await request.send();
+    final responseBody = await response.stream.bytesToString();
+
+    final jsonBody = json.decode(responseBody);
+    if (response.statusCode == 200) {
+      final audioResponse = GroqParser.audioResponseFromJson(jsonBody);
+      final usage =
+          GroqParser.groqUsageFromAudioJson(jsonBody['x_groq']['usage']);
+      final rateLimitInfo =
+          GroqParser.rateLimitInformationFromHeaders(response.headers);
+      return (audioResponse, usage, rateLimitInfo);
+    } else {
+      throw GroqException(
+          statusCode: response.statusCode, error: GroqError.fromJson(jsonBody));
     }
   }
 }