+++ lib/models/groq_audio_response.dart
@@ -1,0 +1,9 @@
+class GroqAudioResponse {
+  final String requestId;
+  final String text;
+
+  GroqAudioResponse({
+    required this.requestId,
+    required this.text,
+  });
+}